wpj package in msalautomationapp
